+++ capturecat-clients/uploader/src/main/java/com/capturecat/client/upload/config/FileUploaderConfig.java
@@ -1,0 +1,21 @@
+package com.capturecat.client.upload.config;
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.s3.S3Client;
+
+@Configuration
+@EnableConfigurationProperties(S3Properties.class)
+public class FileUploaderConfig {
+
+    @Bean
+    @Profile({"dev", "prod"})
+    public S3Client s3Client() {
+        return S3Client.builder()
+                .region(Region.AP_NORTHEAST_2)
+                .build(); // ~/.aws/credentials의 default 프로파일 사용
+    }
+}
