+++ capturecat-core/src/main/java/com/capturecat/CaptureCatApplication.java
@@ -2,10 +2,8 @@ package com.capturecat;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-@ConfigurationPropertiesScan(basePackages = "com.capturecat.core.support")
 @EnableJpaAuditing
 @SpringBootApplication
 public class CaptureCatApplication {
