+++ capturecat-core/src/main/java/com/capturecat/core/config/AppConfig.java
@@ -6,23 +6,12 @@ import org.modelmapper.config.Configuration.AccessLevel;
 import org.modelmapper.convention.MatchingStrategies;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.s3.S3Client;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 @Configuration
 public class AppConfig {
-
-    @Bean
-    @Profile({"dev", "prod"})
-    public S3Client s3Client() {
-        return S3Client.builder()
-                .region(Region.AP_NORTHEAST_2)
-                .build(); // ~/.aws/credentials의 default 프로파일 사용
-    }
 
     @Bean
     public ModelMapper modelMapper() {
