+++ capturecat-core/src/test/java/com/capturecat/core/api/HealthCheckControllerTest.java
@@ -1,6 +1,5 @@
 package com.capturecat.core.api;
 
-import com.capturecat.core.api.HealthCheckController;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
