+++ capturecat-core/src/test/java/com/capturecat/core/service/image/FileUploaderTest.java
@@ -1,5 +1,6 @@
 package com.capturecat.core.service.image;
 
+import com.capturecat.client.upload.api.FileUploader;
 import com.capturecat.core.api.image.dto.ImageMapper;
 import com.capturecat.core.domain.image.ImageRepository;
 import org.junit.jupiter.api.DisplayName;
