+++ settings.gradle
@@ -3,4 +3,4 @@ rootProject.name = 'capture-cat'
 include 'capturecat-core'
 include 'capturecat-support'
 include 'capturecat-tests:api-docs'
-include 'capturecat-clients'
+include 'capturecat-clients:uploader'
